Move visualisation config validation to its own file
