+++ packages/central-server/app/admin/programImporter/programDataElementValidation.js
@@ -1,7 +1,4 @@
-import * as yup from 'yup';
 import config from 'config';
-import { parseOrNull } from '@tamanu/shared/utils/parse-or-null';
-import { isNumberOrFloat } from '../../utils/numbers';
 import { statkey, updateStat } from '../stats';
 import {
   PROGRAM_DATA_ELEMENT_TYPES,
@@ -9,64 +6,7 @@ import {
   SURVEY_TYPES,
   COMPLEX_CORE_DATA_ELEMENT_TYPES,
 } from '@tamanu/constants';
-
-const checkIfWithinGraphRange = (normalRange, graphRange) => {
-  if (isNumberOrFloat(normalRange.min) && normalRange.min < graphRange.min) {
-    return false;
-  }
-  if (isNumberOrFloat(normalRange.max) && normalRange.max > graphRange.max) {
-    return false;
-  }
-  return true;
-};
-
-const validateNormalRangeAsObject = (normalRange, graphRange) => {
-  if (!checkIfWithinGraphRange(normalRange, graphRange)) {
-    throw new Error(
-      `normalRange must be within graphRange, got normalRange: ${JSON.stringify(
-        normalRange,
-      )}, graphRange: ${JSON.stringify(graphRange)}}`,
-    );
-  }
-
-  return true;
-};
-
-const validateNormalRangeAsArray = (normalRange, graphRange) => {
-  for (const normalRangeObject of normalRange) {
-    validateNormalRangeAsObject(normalRangeObject, graphRange);
-  }
-
-  return true;
-};
-
-const validateNormalRange = (normalRange, graphRange) => {
-  if (yup.object().isType(normalRange)) {
-    return validateNormalRangeAsObject(normalRange, graphRange);
-  }
-
-  if (yup.array().isType(normalRange)) {
-    return validateNormalRangeAsArray(normalRange, graphRange);
-  }
-
-  return false;
-};
-
-function validateVisualisationConfig(visualisationConfigString, validationCriteriaString) {
-  const visualisationConfig = parseOrNull(visualisationConfigString);
-  const validationCriteria = parseOrNull(validationCriteriaString);
-
-  if (visualisationConfig) {
-    if (!validationCriteria) {
-      throw new Error('validationCriteria must be specified if visualisationConfig is presented');
-    }
-    if (!validationCriteria.normalRange) {
-      throw new Error('validationCriteria must have normalRange');
-    }
-
-    validateNormalRange(validationCriteria.normalRange, visualisationConfig.yAxis.graphRange);
-  }
-}
+import { validateVisualisationConfig } from './visualisationConfigValidation';
 
 function validateComplexChartCoreConfig(programDataElementRecord, surveyScreenComponentRecord) {
   const { type } = programDataElementRecord.values;
